+++ src/topoppi/gui_app/plot_mixin.py
@@ -1,5 +1,6 @@
 import tkinter as tk
 from tkinter import colorchooser, messagebox
+
 import matplotlib
 
 try: